requirements.txt
# HTTP 요청
requests==2.31.0

# 환경변수 관리
python-dotenv==1.0.0

# 테스팅
pytest==7.4.3
pytest-asyncio==0.21.1

# JSON 처리 및 데이터 검증
pydantic==2.5.0

# 로깅
loguru==0.7.2

# 시간 처리
python-dateutil==2.8.2

# 개발용 유틸리티
black==23.11.0
flake8==6.1.0